Moved Config data model from rag_answers.cli.py to rag_answers.data_models.py and deleted now redundant EvaluationConfig after updating Config with .metric_instances() method; updated govuk_chat_evaluation/rag_answers/data_models.py govuk_chat_evaluation/rag_answers/cli.py govuk_chat_evaluation/rag_answers/evaluate.py accordingly.
