+++ govuk_chat_evaluation/rag_answers/cli.py
@@ -3,23 +3,12 @@ from pathlib import Path
 from typing import cast
 
 import click
-from pydantic import model_validator
 
-from ..config import BaseConfig, apply_click_options_to_command, config_from_cli_args
+from ..config import apply_click_options_to_command, config_from_cli_args
 from ..file_system import create_output_directory, write_config_file_for_reuse
 from .evaluate import evaluate_and_output_results
 from .generate import generate_and_write_dataset
-
-
-class Config(BaseConfig):
-    what: BaseConfig.GenericFields.what
-    generate: BaseConfig.GenericFields.generate
-    provider: BaseConfig.GenericFields.provider_openai_or_claude
-    input_path: BaseConfig.GenericFields.input_path
-
-    @model_validator(mode="after")
-    def run_validatons(self):
-        return self._validate_fields_required_for_generate("provider")
+from .data_models import Config
 
 
 @click.command(name="rag_answers")
@@ -48,6 +37,6 @@ def main(**cli_args):
     else:
         evaluate_path = config.input_path
 
-    evaluate_and_output_results(output_dir, evaluate_path)
+    evaluate_and_output_results(output_dir, evaluate_path, config)
 
     write_config_file_for_reuse(output_dir, config)
